Move unicode converter code to its own module
